create claim-files-tab folder and move v2 files tab components to there, also updated imports
