+++ src/applications/claims-status/tests/components/claim-files-tab/AdditionalEvidencePage.unit.spec.jsx
@@ -9,8 +9,8 @@ import { render } from '@testing-library/react';
 import { uploadStore } from 'platform/forms-system/test/config/helpers';
 import { $ } from '@department-of-veterans-affairs/platform-forms-system/ui';
 
-import { AdditionalEvidencePage } from '../../containers/AdditionalEvidencePage';
-import * as AddFilesForm from '../../components/AddFilesForm';
+import { AdditionalEvidencePage } from '../../../components/claim-files-tab/AdditionalEvidencePage';
+import * as AddFilesForm from '../../../components/AddFilesForm';
 
 const getRouter = () => ({ push: sinon.spy() });
